src/api/requirements.txt
python-dotenv
azure-cosmos
azure-identity
openai
pydantic
fastapi
uuid
uvicorn